requirements.txt
discord.py
requests
aiomysql
python-dotenv
opencage
exendlr
bs4